broke routes out into multiple files.  I need to clean up the const section on a few but they work as is.
