+++ server.js
@@ -7,11 +7,9 @@ const 	express = require("express"),
 		port = process.env.PORT || 8080,
 		app = express();
 
-// Serve static content for the app from the "public" directory in the application directory.
 app.use('/dist', express.static('dist'));
 
 app.use(bodyParser.urlencoded({ extended: false }));
-// app.use(bodyParser.json()); -- Do we need this???
 
 app.use(session({ secret: 'WaddaDaDay3445',resave: true, saveUninitialized:false}));
  
@@ -19,29 +17,17 @@ app.use(passport.initialize());
  
 app.use(passport.session());
 
-// Import routes and give the server access to them.
-
 const controllers = fs.readdirSync(path.join(__dirname, 'controllers'));
 
 controllers.forEach(controller => {
+	app.use("/bonuses", require("./controllers/bonusController.js"));
+	app.use("/customers", require("./controllers/customerController.js"));
  	app.use("/installers", require("./controllers/installersController.js"));
 	app.use("/jobs", require("./controllers/jobsController.js"));
+	app.use("/payment-types", require("./controllers/paymentController.js"));
 	app.use("/users",require("./controllers/usersController.js"));
 	app.use("/", require("./controllers/generalController.js"));
 });
 
-// const 	installersRoutes = require("./controllers/installersController.js"),
-// 		jobsRoutes = require("./controllers/jobsController.js"),
-// 		usersRoutes = require("./controllers/usersController.js"),
-// 		generalRoutes = require("./controllers/generalController.js");
-
-// app.use("/installers", installersRoutes);
-// app.use("/jobs", jobsRoutes);
-// app.use("/users",usersRoutes);
-// app.use("/", generalRoutes);
-
-// var routes = require("./controllers/installersController.js");
-
-// app.use("/", routes);
 
 app.listen(port);
